Implementing #170 and #178
