+++ CHANGES
@@ -7,6 +7,9 @@ Version 1.1.0
 -------------
 
 - Feature: In-memory plug-ins (see #140)
+- Enhancement: Logic unified between pyblish.util and pyblish.cli
+- Bugfix: Order now works with pyblish.util and pyblish.cli
+- API: Added pyblish.api.plugins_by_instance()
 - API: New defaults for `hosts` and `families` of plug-ins. (see #176)
 - API: Added pyblish.api.register_plugin()
 - API: Added pyblish.api.deregister_plugin()
